% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch-package.R
\docType{package}
\name{ch-package}
\alias{ch}
\alias{ch-package}
\title{ch: About Chemical Methods}
\description{
Some  small functions. Some functions to chem.
}
\author{
\strong{Maintainer}: Yi Tsia \email{tsiamut@gmail.com}

}
\keyword{internal}
