% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_cl.R
\name{console_cl}
\alias{console_cl}
\title{Remove > and + from the run area and add # to the run result}
\usage{
console_cl(preffix = "#>")
}
\arguments{
\item{preffix}{The preffix for code.}
}
\description{
Remove '>' and '+' from the run area and add # to
the run result.
}
\details{
Remove > and + from the run area and add # to the run result
}
\author{
Chai
}
