% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2vcf.R
\name{csv2vcf}
\alias{csv2vcf}
\title{about csv2vcf}
\usage{
csv2vcf(csv_file, vcf_file, header = F)
}
\arguments{
\item{csv_file}{the csv file to choose.}

\item{vcf_file}{save the vcf file.}

\item{header}{for more see \code{\link[utils]{read.csv}}}
}
\description{
The easiest way to generate vcf files.
}
