% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_multicol.R
\name{df_multicol}
\alias{df_multicol}
\title{The realization of multi-column repeated display of three-line
table data generation.}
\usage{
df_multicol(df, n)
}
\arguments{
\item{df}{data.frame}

\item{n}{the numbers}
}
\description{
The realization of multi-column repeated display of three-line
table data generation.
}
\examples{
\dontrun{
d1 <- data.frame(x1 = 1:26,
    y = letters)
d1 \%>\%  df_multicol(2) \%>\%  xtable
d1 \%>\%  df_multicol(3) \%>\%  xtable
d1 \%>\%  df_multicol(4) \%>\%  xtable
}
}
\author{
Chai
}
