% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down.file.R
\name{down_files}
\alias{down_files}
\title{download files}
\usage{
down_files(url, numbers, filename)
}
\arguments{
\item{url}{a website}

\item{numbers}{the numbers need to count}

\item{filename}{a file to save, that is a pdf}
}
\description{
To download files form the url
}
\details{
download files
}
\author{
tsia
}
