% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showruler.R
\name{draw_ruler}
\alias{draw_ruler}
\title{some examples about ggplot2}
\usage{
draw_ruler(len = 5)
}
\arguments{
\item{len}{the length of ruler}
}
\description{
different examples
}
\details{
some examples about ggplot2
}
\author{
tsia
}
