% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ground-State.R
\name{ground_state}
\alias{ground_state}
\alias{state_1}
\alias{state_2}
\title{Ground-State Spectroscopic Term}
\usage{
ground_state(x)

state_1(l, n)

state_2(x)
}
\arguments{
\item{x}{'p2','p3','d2','d5','f2',...}

\item{l}{the l:0,1,2,3,4,5,...}

\item{n}{the number of electrons.}
}
\description{
Use such a function to calculate the spectrum item,
and  it can show the number of the spectrum item.
}
\examples{
\dontrun{
  ground_state(1,2)
  ground_state(3,2)
 state_2('p2')
 state_2('d3')
 }
}
\author{
Chai
}
