% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latex_fmt}
\alias{latex_fmt}
\alias{lat_fmt}
\title{tex format and $$...$$}
\usage{
latex_fmt(x)

lat_fmt(x)
}
\arguments{
\item{x}{Ryacas object}
}
\description{
tex output to clipboard

output to console
}
\details{
tex format for symbolic computation output
}
\examples{
\dontrun{
p_load(Ryacas)
r <- ysym('r')
B <- ysym('B')
a <- ysym('a')
x1 <- A*B*r*exp(-B*r/2)* cos(a)
deriv(x1,'r') \%>\%  simplify \%>\% latex
deriv(x1,'r') \%>\%  simplify \%>\% la
deriv(x1,'r') \%>\%  simplify \%>\% lat
}
}
\author{
Chai
}
