% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoa.R
\name{monoa}
\alias{monoa}
\title{Calculate The pH}
\usage{
monoa(ka, c, digits = 2)
}
\arguments{
\item{ka}{ionization constant}

\item{c}{concentration}

\item{digits}{digit}
}
\description{
Calculate the pH of weak acids
}
\examples{
monoa(1.35e-5,0.01)
}
