% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoa.R
\name{monob}
\alias{monob}
\title{Calculate The pH}
\usage{
monob(ka, c, digits = 2)
}
\arguments{
\item{ka}{ionization constant}

\item{c}{concentration}

\item{digits}{digit}
}
\description{
Calculate the pH of weak bases.
}
\examples{
monob(1.35e-5,0.01)
}
