% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o.R
\name{op}
\alias{op}
\title{open a file on Windows System}
\usage{
op(...)
}
\arguments{
\item{...}{a file to open}
}
\description{
open a file on Windows System
}
\details{
open a file on Windows System

Only in current path, the file doesn't  need
to add the quotation marks. When adding the path, you must
add the quotes.
}
\author{
tsia
}
