% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orthogonal_experimental_design.R
\name{ortho_design}
\alias{ortho_design}
\title{Orthogonal experimental design}
\usage{
ortho_design(df, n_levels, digits = 3)
}
\arguments{
\item{df}{data.frame}

\item{n_levels}{numbers of levels}

\item{digits}{digits of output}
}
\description{
Orthogonal experimental design
}
