% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table.R
\name{plot_table}
\alias{plot_table}
\title{use ggplot2 to plot a table}
\usage{
plot_table(Str, ncol, byrow = T)
}
\arguments{
\item{Str}{some expressions}

\item{ncol}{the number f col}

\item{byrow}{logical}
}
\description{
draw table in ggplot2
}
\details{
some examples about ggplot2
}
\author{
Chai
}
