% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Q_test.R
\name{Q.test}
\alias{Q.test}
\title{Q test}
\usage{
Q.test(x, digits = 2)
}
\arguments{
\item{x}{vector}

\item{digits}{the number of digits}
}
\description{
Complete the Q test
}
\details{
Q test
}
\examples{
c(0.001,0.023,0.015,0.018,0.023,0.034,0.025) -> a2
Q.test(a2)
}
