% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.R
\name{rd}
\alias{rd}
\title{relative  deviation}
\usage{
rd(x, digits = 4)
}
\arguments{
\item{x}{a vector}

\item{digits}{digit}
}
\description{
Calculate relative  deviation
}
\details{
relative  deviation
}
\examples{
 a1 = c(0.001,0.023,0.015)
 rd(a1)
}
\author{
tsia
}
