% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_mean.R
\name{rd_mean}
\alias{rd_mean}
\title{relative mean deviation}
\usage{
rd_mean(x, digits = 4)
}
\arguments{
\item{x}{vector}

\item{digits}{digit}
}
\description{
Calculate relative mean deviation
}
\details{
Relative mean deviation
}
\examples{
rd_mean(c(0.001,0.023,0.01))
}
