% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readtxt.R
\name{read.txt}
\alias{read.txt}
\title{read the text to data.frame}
\usage{
read.txt(text, header = T, ...)
}
\arguments{
\item{text}{strings}

\item{header}{logical value}

\item{...}{for read.table}
}
\description{
read the strings and transform to
the data.frame
}
\author{
Chai
}
