% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round.R
\name{Round}
\alias{Round}
\alias{round_1}
\alias{Round2}
\title{To achieve the standard sense of rounding and
rounding off to five in double.}
\usage{
Round(x, n = 0)

round_1(x, n)

Round2(x, n)
}
\arguments{
\item{x}{vector or matrix}

\item{n}{digits}
}
\description{
Round : rounding off to five in double
round_1 and Round2: to achieve the standard sense of rounding
}
\author{
Chai
}
