% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsd.R
\name{rsd}
\alias{rsd}
\title{relative standard deviation}
\usage{
rsd(x, digits = 4)
}
\arguments{
\item{x}{vector}

\item{digits}{the finally digits}
}
\description{
Calculate relative standard deviation
}
\details{
relative standard deviation
}
\author{
tsia
}
