% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanstr.R
\name{scan.str}
\alias{scan.str}
\title{Read string into a vector}
\usage{
scan.str(string)
}
\arguments{
\item{string}{a string that number is separated by ' '.}
}
\value{
A vector that contains numbers.
}
\description{
Read data into a vector  from a string.
}
\examples{
m <- "12 23 45 78 90 89 97"
scan.str(m)
}
\author{
Chai
}
