% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_color.R
\name{show_color}
\alias{show_color}
\title{some examples about ggplot2}
\usage{
show_color(
  colors,
  ncol,
  byrow = T,
  label = F,
  number = F,
  size = 1,
  border = "black"
)
}
\arguments{
\item{colors}{string about colors}

\item{ncol}{the number f col}

\item{byrow}{logical}

\item{label}{logical}

\item{number}{logical}

\item{size}{the size of label}

\item{border}{about the border}
}
\description{
different examples
}
\details{
some examples about ggplot2
}
\author{
tsia
}
