% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showruler.R
\name{show_ruler}
\alias{show_ruler}
\title{Use ggplot to draw ruler}
\usage{
show_ruler(len = 5)
}
\arguments{
\item{len}{the length of ruler}
}
\value{
A ggplot2 object.
}
\description{
You can draw a ruler that uses ggplot2.
Also It's funny and you can get a lot of methods from
this function.
}
\author{
Chai
}
