% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvef.R
\name{solvef}
\alias{solvef}
\title{linear equations}
\usage{
solvef(a, b)
}
\arguments{
\item{a}{the first equation}

\item{b}{the second equation}
}
\description{
solve linear equations
}
\details{
Solve a system of binary linear equations
}
\examples{
solvef(c(1,2,3),c(2.5,3,6))
}
