% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{stat}
\alias{stat}
\title{state}
\usage{
stat(x)
}
\arguments{
\item{x}{Like: d5,p2,p3,f3}
}
\value{
the state content.
}
\description{
About state, new function!
}
\examples{
stat("d5")
stat("p2")
stat("p3")
stat("f3")
}
