% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ryacas-all.R
\name{sym2poly}
\alias{sym2poly}
\alias{sym2coef}
\title{Extracting coefficients from polynomials}
\usage{
sym2poly(x, var = "x")

sym2coef(x, var = "x")
}
\arguments{
\item{x}{The polynomials,for examples,
'3\emph{x^2 + 6}x^6 + 2 + 25*x'.}

\item{var, }{The var from polynomials, for examples,
the var of '3\emph{x^2 + x^6 + x}8 + x*5^2' is 'x'.}
}
\value{
sym2poly() returns Coefficients and the roots.
sym2coef() only returns coefficients.
}
\description{
sym2ploy can extract coefficients from
polynomials and gives the roots of polynomials.
The roots is  calculated from \code{\link[base]{polyroot}} and
the \code{\link[pracma]{polyroots}} function.
}
\examples{
sym2poly("3*x^2 + x^5 + x*8")
}
\author{
Chai
}
