% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_cl.R
\name{console_cl}
\alias{console_cl}
\title{Remove '>' and '+' from the console and
add '#' to the run result.}
\usage{
console_cl(prefix = "#>")
}
\arguments{
\item{prefix}{The prefix for code.The default is '#>'.
You can edit it according to your own preference, but
'#' should be  the first character.}
}
\value{
the result of the run is saved to the clipboard.
}
\description{
First you need to copy the console area to the clipboard,
then run the console_cl() function to add a comment to the line
where the output is, and to cancel the > on the original line.
Finally, the result of the run is saved to the clipboard.
}
\author{
Chai
}
