% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2vcf.R
\name{csv2vcf}
\alias{csv2vcf}
\title{about csv2vcf}
\usage{
csv2vcf(csv_file, vcf_file, header = FALSE)
}
\arguments{
\item{csv_file}{The csv file contains
names and  phone numbers.
The style is like this:
Joey  18100
Hans  12788
Tim   34689
The first column is the name,
the second column is the phone number corresponding to each person.
The above is an example, and it is not true personal information summary.}

\item{vcf_file}{The vcf file to create.}

\item{header}{For more see \code{\link[utils]{read.csv}},
the default is FALSE.}
}
\value{
NULL. t will be saved in a file with the suffix vcf.
}
\description{
A simple method to  generate vcf file.
}
