% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2vcf.R
\name{csv2vcf}
\alias{csv2vcf}
\title{about csv2vcf}
\usage{
csv2vcf(csv_file, vcf_file, header = F)
}
\arguments{
\item{csv_file}{The csv file contains
names and  phone numbers.}

\item{vcf_file}{The vcf file to create.}

\item{header}{For more see \code{\link[utils]{read.csv}}}
}
\description{
A simple method to generate VCF generate vcf files.
}
