% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpng.R
\name{ggpng}
\alias{ggpng}
\alias{ggPNG}
\title{Create a picture with a transparent background.}
\usage{
ggpng(x, dpi = 600, ...)

ggPNG(x, p, dpi = 600, ...)
}
\arguments{
\item{x}{A file name that does not have a suffix.}

\item{dpi}{The default dpi is 600.
You can enhance the dpi value to produce a higher resolution PNG file.}

\item{...}{see :\link{ggsave}}

\item{p}{ggplot2 object}
}
\value{
You will get a PNG file with the result drawn by ggplot2.
}
\description{
Use the ImageMagick command line to convert the PDF saved
by ggplot2 to PNG format with a transparent background and to
set the resolution.
}
\details{
You need to install ImageMagick!
Please check if the ImageMagick is added to the environment variable.

this ggplot2 object will automatically add a theme with a
transparent background.
}
\author{
Chai
}
