% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ground-State.R
\name{ground_state}
\alias{ground_state}
\alias{state_1}
\alias{state_2}
\title{Ground-State spectral term}
\usage{
ground_state(x)

state_1(l, n)

state_2(x)
}
\arguments{
\item{x}{'p2','p3','d2','d5','f2',...}

\item{l}{the l:0,1,2,3,4,5,...}

\item{n}{the number of electrons.}
}
\value{
It is a display of the results of the calculation of
the spectral term. For more explanation, please
refer to the structural chemistry.
The ground_state() function will tell you
the ground state that spectral term of  equal electrons.
The state_1() and state_2() will tell you the
spectral term of equal electrons.
}
\description{
Use such a function to calculate the spectral term,
and  it can show the number of the spectral term.
}
\examples{
ground_state("p2")
ground_state("f3")
state_2("p2")
state_2("d3")
}
\references{
The method of  state_1() and state_2() function is from:
DOI:10.14159/j.cnki.0441-3776.1985.11.020
And the url is:
\url{https://t.cnki.net/kcms/detail?v=3uoqIhG8C44YLTlOAiTRKqd0WnNPv0wTDjtDUwHroNz8ZoQZVLjnVKa9t-3R3F8t9DSW1fOAkwcAF8JZi9EnOU8ccrbY3bNC&uniplatform=NZKPT}
You can get more details from this essay.
}
\author{
Chai
}
