% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ryacas-all.R
\name{lat_fmt}
\alias{lat_fmt}
\alias{latex_fmt}
\title{tex format and $$...$$}
\usage{
lat_fmt(x)

latex_fmt(x)
}
\arguments{
\item{x}{symbol object,for more see \code{\link[Ryacas]{ysym}}.}
}
\value{
lat_fmt() will output to Console.
latex_fmt() will output to clipboard.
}
\description{
output to Console and clipboard.
You can better check the correctness of the output.
}
\author{
Chai
}
