% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoa.R
\name{monoa}
\alias{monoa}
\alias{monob}
\alias{mono}
\title{Calculate The pH of weak acid/base}
\usage{
monoa(ka, c, digits = 2)

monob(ka, c, digits = 2)

mono(ka, c, digits = 2, acid = TRUE, kw = 1e-14)
}
\arguments{
\item{ka}{ionization constant.}

\item{c}{concentration.}

\item{digits}{digit of the output.}

\item{acid}{if TRUE, it is equivalent
to monoa function; if FALSE, it is
equivalent to monob function.}

\item{kw}{the default is 1e-14}
}
\value{
monoa() will return the pH of weak acid,
the monob() will return the pH of weak base.
And you can also use the mono() function to
replace the monoa()  function and monob() function.
}
\description{
Calculate the pH of weak acid or base.
}
\examples{
monoa(1.4 - 6, 2.35e-2)
monoa(2.78e-8, 0.01)
monob(1.35e-5, 0.01)
monob(2.4 - 6, 1e-4)
}
