% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoa.R
\name{monoa}
\alias{monoa}
\alias{monob}
\alias{mono}
\title{Calculate The pH}
\usage{
monoa(ka, c, digits = 2)

monob(ka, c, digits = 2)

mono(ka, c, digits = 2, acid = T)
}
\arguments{
\item{ka}{ionization constant}

\item{c}{concentration}

\item{digits}{digit}

\item{acid}{if TRUE, it is equivalent
to monoa function; if FALSE, it is
equivalent to monob function.}
}
\description{
Calculate the pH of weak acids

Calculate the pH of weak bases.
}
\examples{
monoa(1.35e-5,0.01)
monob(1.35e-5,0.01)
}
\author{
Chai
}
