% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table.R
\name{plot_table}
\alias{plot_table}
\title{Use ggplot2 to plot a table}
\usage{
plot_table(Str, ncol, byrow = TRUE)
}
\arguments{
\item{Str}{some expressions}

\item{ncol}{the number f col}

\item{byrow}{logical, the default is}
}
\value{
It will output a ggplot object that
contains a table.
}
\description{
You can use plot_table to  draw table in ggplot2,
but it only applies to expressions(see \code{\link[base]{expression}}).
For more information , you can see \code{\link[grDevices]{plotmath}}.
But it's a ggplot2 object!
}
\examples{
a1 <- c(
  "x \%*\% y", "x \%/\% y", "alpha", "sigma", "beta",
  "x == y", "frac(x,y)", "x \%up\% y", "hat(x)",
  "symbol(a)", "underline(x)"
)
plot_table(a1, 2)
plot_table(a1, 3)
}
\author{
Chai
}
