% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-some.R
\name{plot_table}
\alias{plot_table}
\title{some examples about ggplot2}
\usage{
plot_table(x, ncol, byrow = T)
}
\arguments{
\item{x}{some expressions}

\item{ncol}{the number f col}

\item{byrow}{logical}
}
\description{
different examples
}
\details{
some examples about ggplot2
}
\author{
tsia
}
