% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readtxt.R
\name{read.txt}
\alias{read.txt}
\title{Read the text to data.frame}
\usage{
read.txt(text, header = TRUE, ...)
}
\arguments{
\item{text}{strings}

\item{header}{logical value}

\item{...}{for more see \code{\link[utils]{read.table}}}
}
\value{
A data.frame
}
\description{
Read the strings and transform to
the data.frame.
}
\author{
Chai
}
