% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_color.R
\name{show_color}
\alias{show_color}
\title{An easy way to show  colors in ggplot2}
\usage{
show_color(
  colors,
  ncol,
  byrow = TRUE,
  label = FALSE,
  number = FALSE,
  size = 1,
  border = "black"
)
}
\arguments{
\item{colors}{string about colors}

\item{ncol}{the number f col}

\item{byrow}{logical}

\item{label}{logical}

\item{number}{logical, the default is}

\item{size}{the size of label, the default is 1.}

\item{border}{The color of  border}
}
\description{
the same function can see
\code{\link[scales]{show_col}}, but it is a
ggplot2 object. You can use it like the show_col()
function in scales package, but it can save by ggsave() function.
}
\author{
Chai
}
