% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_color.R
\name{show_color}
\alias{show_color}
\title{An easy way to show in ggplot2}
\usage{
show_color(
  colors,
  ncol,
  byrow = T,
  label = F,
  number = F,
  size = 1,
  border = "black"
)
}
\arguments{
\item{colors}{string about colors}

\item{ncol}{the number f col}

\item{byrow}{logical}

\item{label}{logical}

\item{number}{logical}

\item{size}{the size of label}

\item{border}{about the border}
}
\description{
the same function can see
\code{\link[scales]{show_col}}, but it is a
ggplot2 object.
}
\author{
Chai
}
