% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symb.R
\name{symb}
\alias{symb}
\title{Generating multiple symbols at once}
\usage{
symb(..., envir = parent.frame(), quite = FALSE)
}
\arguments{
\item{...}{The multiple vectors.}

\item{envir}{The environment.}

\item{quite}{If FALSE, it will show the message in the end.}
}
\value{
The multiple symbols.
}
\description{
It may be faster than using \code{\link[Ryacas]{ysym}}.
}
\examples{
\donttest{
library(Ryacas)
symb(x, y, z)
str(x)
}
}
\author{
Chai
}
